*: reorganize the project layout
