+++ .gitignore
@@ -1,3 +1,9 @@
+bin/
+*.tar
+*.gz
+*.xz
+
+# ---> Go
 # Binaries for programs and plugins
 *.exe
 *.exe~
@@ -5,11 +11,10 @@
 *.so
 *.dylib
 
-# Test binary, built with `go test -c`
+# Test binary, build with `go test -c`
 *.test
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 
-# Dependency directories (remove the comment below to include it)
-# vendor/
+/.idea/
