+++ README.md
@@ -1,5 +1,17 @@
-# TiUP Demo
-This is prototype demos of TiUP and TiAdm.
+# TiUP
 
-# Note
-This is only demos of prototype designs, it shows how things are working, but not doing any operations for real.
+`tiup` is a tool to download and install TiDB components.
+
+# Usage
+After installing `tiup`, you can use it to install binaries of TiDB components.
+
+```
+tiup [FLAGS]
+  -V, --version   Show tiup version and quit
+
+tiup COMMAND [FLAGS]
+  show        Show available TiDB components
+        --all    Show all available components and versions (refresh online).
+  install     Install component(s) of specific version
+  uninstall   Remove installed component(s)
+```
