+++ cmd/base.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/AstroProfundis/tiup-demo/tiup/pkg/version"
+	"github.com/c4pt0r/tiup/pkg/version"
 	"github.com/spf13/cobra"
 )
