+++ cmd/component.go
