+++ cmd/launch.go
@@ -6,8 +6,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/AstroProfundis/tiup-demo/tiup/pkg/meta"
-	"github.com/AstroProfundis/tiup-demo/tiup/pkg/utils"
+	"github.com/c4pt0r/tiup/pkg/meta"
+	"github.com/c4pt0r/tiup/pkg/utils"
 	"github.com/spf13/cobra"
 )
