--- tiup/.gitignore
@@ -1,19 +1,0 @@
-bin/
-*.tar
-*.gz
-*.xz
-
-# ---> Go
-# Binaries for programs and plugins
-*.exe
-*.exe~
-*.dll
-*.so
-*.dylib
-
-# Test binary, build with `go test -c`
-*.test
-
-# Output of the go coverage tool, specifically when used with LiteIDE
-*.out
-
