--- tiup/README.md
@@ -1,17 +1,0 @@
-# TiUP
-
-`tiup` is a tool to download and install TiDB components.
-
-# Usage
-After installing `tiup`, you can use it to install binaries of TiDB components.
-
-```
-tiup [FLAGS]
-  -V, --version   Show tiup version and quit
-
-tiup COMMAND [FLAGS]
-  show        Show available TiDB components
-        --all    Show all available components and versions (refresh online).
-  install     Install component(s) of specific version
-  uninstall   Remove installed component(s)
-```
